GameResult 와 DealerResult 분리
