+++ Blackjack/Model/GameResult.swift
@@ -15,12 +15,6 @@ struct GameResult {
         case lose
     }
     
-    struct DealerOutcome {
-        let winningCount: Int
-        let drawCount: Int
-        let loseCount: Int
-    }
-    
     private let dealer: Dealer
     private let gamers: [Gamer]
     private var winningPoint: Int {
@@ -31,36 +25,6 @@ struct GameResult {
             }
             .map { $0.totalPoint }.max() ?? 0
         return winningPoint
-    }
-    
-    var dealerOutcome: DealerOutcome {
-        
-        guard !dealer.isBurst else {
-            return DealerOutcome(winningCount: 0,
-                                 drawCount: 0,
-                                 loseCount: gamers.count)
-        }
-        
-        let winningCount: Int = gamers.filter { gamer in
-            let isGamerBurst: Bool = gamer.isBurst
-            let isHigherPointDealer: Bool = gamer.totalPoint < dealer.totalPoint
-            return isGamerBurst || isHigherPointDealer
-        }.count
-        
-        let drawCount: Int = gamers.filter { gamer in
-            let isSamePointDealer: Bool = gamer.totalPoint == dealer.totalPoint
-            return isSamePointDealer
-        }.count
-        
-        let loseCount: Int = gamers.filter { gamer in
-            let isGamerBurst: Bool = gamer.isBurst
-            let isLowerPointDealer: Bool = gamer.totalPoint > dealer.totalPoint
-            return isLowerPointDealer && !isGamerBurst
-        }.count
-        
-        return DealerOutcome(winningCount: winningCount,
-                             drawCount: drawCount,
-                             loseCount: loseCount)
     }
     
     init(dealer: Dealer, gamers: [Gamer]) {
